macro: Restructure `ink_macro::sol` module
