+++ crates/ink/macro/src/lib.rs
@@ -1602,6 +1602,42 @@ synstructure::decl_derive!(
     storage::storage_layout_derive
 );
 
+/// Derive the re-exported traits `ink::scale::Encode`, `ink::scale::Decode` and
+/// `ink::scale_info::TypeInfo`. It enables using the built in derive macros for these
+/// traits without depending directly on the `parity-scale-codec` and `scale-info` crates.
+///
+/// # Options
+///   - `Encode`: derives `ink::scale::Encode`
+///   - `Decode`: derives `ink::scale::Decode`
+///   - `TypeInfo`: derives `ink::scale_info::TypeInfo`
+///
+/// # Examples
+///
+/// ```
+/// #[ink::scale_derive(Encode, Decode, TypeInfo)]
+/// pub enum Error {}
+/// ```
+/// This is a convenience macro that expands to include the additional `crate` attributes
+/// required for the path of the re-exported crates.
+///
+/// ```
+/// #[derive(::ink::scale::Encode, ::ink::scale::Decode)]
+/// #[codec(crate = ::ink::scale)]
+/// #[cfg_attr(
+///   feature = "std",
+///   derive(::scale_info::TypeInfo),
+///   scale_info(crate = ::ink::scale_info)
+/// )]
+/// pub enum Error {}
+/// ```
+#[proc_macro_attribute]
+pub fn scale_derive(attr: TokenStream, item: TokenStream) -> TokenStream {
+    match scale::derive(attr.into(), item.into()) {
+        Ok(output) => output.into(),
+        Err(err) => err.to_compile_error().into(),
+    }
+}
+
 synstructure::decl_derive!(
     [SolErrorDecode] =>
     /// Derives an implementation of `ink::sol::SolErrorDecode`
@@ -1683,42 +1719,6 @@ synstructure::decl_derive!(
     /// ```
     sol::sol_error_encode_derive
 );
-
-/// Derive the re-exported traits `ink::scale::Encode`, `ink::scale::Decode` and
-/// `ink::scale_info::TypeInfo`. It enables using the built in derive macros for these
-/// traits without depending directly on the `parity-scale-codec` and `scale-info` crates.
-///
-/// # Options
-///   - `Encode`: derives `ink::scale::Encode`
-///   - `Decode`: derives `ink::scale::Decode`
-///   - `TypeInfo`: derives `ink::scale_info::TypeInfo`
-///
-/// # Examples
-///
-/// ```
-/// #[ink::scale_derive(Encode, Decode, TypeInfo)]
-/// pub enum Error {}
-/// ```
-/// This is a convenience macro that expands to include the additional `crate` attributes
-/// required for the path of the re-exported crates.
-///
-/// ```
-/// #[derive(::ink::scale::Encode, ::ink::scale::Decode)]
-/// #[codec(crate = ::ink::scale)]
-/// #[cfg_attr(
-///   feature = "std",
-///   derive(::scale_info::TypeInfo),
-///   scale_info(crate = ::ink::scale_info)
-/// )]
-/// pub enum Error {}
-/// ```
-#[proc_macro_attribute]
-pub fn scale_derive(attr: TokenStream, item: TokenStream) -> TokenStream {
-    match scale::derive(attr.into(), item.into()) {
-        Ok(output) => output.into(),
-        Err(err) => err.to_compile_error().into(),
-    }
-}
 
 #[cfg(test)]
 pub use contract::generate_or_err;
